+++ lib/sg_pt_solaris.c
@@ -175,7 +175,7 @@ void set_scsi_pt_packet_id(struct sg_pt_base * vp, int pack_id)
     pack_id = pack_id;          /* ignore and suppress warning */
 }
 
-void set_scsi_pt_tag(struct sg_pt_base * vp, unsigned long long tag)
+void set_scsi_pt_tag(struct sg_pt_base * vp, uint64_t tag)
 {
     // struct sg_pt_solaris_scsi * ptp = &vp->impl;
